+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Fixed
 * Incorrect order of tag/key/value history
+* Error when running emacs with `-q`
 
 ## [0.6.3] - 2021-01-19
 ### Fixed